backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
Flask-Limiter==3.5.0